migrate styles and f related iles to entityDetail folder
