+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailHeaderRow/EntityDetailHeaderRow.tsx
