+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailHeaderRow/EntityDetailHeaderRowStyles.tsx
